refactor: reorganize tests by source module
Rename test files from PR-batch naming to module-based naming:
- test_client.py — all ComfyUIClient method tests
- test_error_hints.py — all error pattern matching tests
- test_nodes.py — nodes command helper tests
- test_update_check.py — unchanged
Add tests/__init__.py for test discovery.
